refactor(tests): extract shared targets fixtures to _targets_helpers.py
Move _settings_env, _isolated_jwks_cache, _empty_connector_registry,
_build_app, _admin_token, and _operator_token into a shared helper
module to eliminate duplication across test_api_v1_targets.py and
test_targets_audit_integration.py (fixes SonarCloud quality gate).
